+++ CHANGELOG.md
@@ -94,7 +94,7 @@
   - Updated to bootstrap 4.0.0.beta2
   - Updated to carrierwave 1.2.1
   - Updated to haml 5.0.4
-  - Updated to kaminari 1.1.0
+  - Updated to kaminari 1.1.1
   - Added bootsnap 1.1.5
 
 ### Tests